% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agaricus_bisporus.R
\docType{data}
\name{Agaricus_bisporus.PhyloMap}
\alias{Agaricus_bisporus.PhyloMap}
\title{Phylomap of Agaricus bisporus}
\format{
A tibble with 11,204 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Year price was recorded}
\item{GeneID}{chr GeneID annotation from UniProt}
}
}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-023-02895-z}
}
\usage{
Agaricus_bisporus.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on refence protein sequences from Uniprot proteomes.
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{ultra-sensitive mode}.
}
\keyword{datasets}
