% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amphimedon_queenslandica_ENSEMBL_Aqu2.1.R
\docType{data}
\name{Amphimedon_queenslandica_ENSEMBL_Aqu2.1.PhyloMap}
\alias{Amphimedon_queenslandica_ENSEMBL_Aqu2.1.PhyloMap}
\title{Phylomap of Amphimedon queenslandica ENSEMBL Aqu2.1}
\format{
A tibble with 43,615 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{ProteinID annotation from ENSEMBL}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Amphimedon_queenslandica_ENSEMBL_Aqu2.1.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from ENSEMBL proteomes (Aqu2.1).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
