% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Armillaria_ostoyae.R
\docType{data}
\name{Armillaria_ostoyae.PhyloMap}
\alias{Armillaria_ostoyae.PhyloMap}
\title{Phylomap of Armillaria ostoyae}
\format{
A tibble with 22,705 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from JGI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Armillaria_ostoyae.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from JGI proteomes (Reference sequence file: p3_i2_t47428_Arm_ostoy_v2.scaf).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
