% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auriculariopsis_ampla.R
\docType{data}
\name{Auriculariopsis_ampla.PhyloMap}
\alias{Auriculariopsis_ampla.PhyloMap}
\title{Phylomap of Auriculariopsis ampla}
\format{
A tibble with 26,484 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from JGI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Auriculariopsis_ampla.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from JGI proteomes (Auriculariopsis ampla NL-1724 v1.0 (Auramp1)).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
