% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bacillus_subtilis.R
\docType{data}
\name{Bacillus_subtilis.PhyloMap}
\alias{Bacillus_subtilis.PhyloMap}
\title{Phylomap of Bacillus subtilis}
\format{
A tibble with 4,333 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{proteinID annotation from NCBI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Bacillus_subtilis.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from NCBI protein coding genes (https://www.ncbi.nlm.nih.gov/assembly/GCF_002055965.1/).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
