% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Borreliella_burgdorferi.R
\docType{data}
\name{Borreliella_burgdorferi.PhyloMap}
\alias{Borreliella_burgdorferi.PhyloMap}
\title{Phylomap of Borreliella burgdorferi}
\format{
A tibble with 1,367 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{proteinID annotation from NCBI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Borreliella_burgdorferi.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from NCBI protein coding genes (https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_000008685.2/).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
