% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brassica_juncea.R
\docType{data}
\name{Brassica_juncea.PhyloMap}
\alias{Brassica_juncea.PhyloMap}
\title{Phylomap of Brassica juncea}
\format{
A tibble with 79,644 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from brassicadb}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Brassica_juncea.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from brassicadb (Braju_tum_V1.5).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
