% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brassica_napus.R
\docType{data}
\name{Brassica_napus.PhyloMap}
\alias{Brassica_napus.PhyloMap}
\title{Phylomap of Brassica napus}
\format{
A tibble with 101,040 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from brassicadb}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Brassica_napus.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from brassicadb (Brana_Dar_V5).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
