% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brassica_nigra.R
\docType{data}
\name{Brassica_nigra.PhyloMap}
\alias{Brassica_nigra.PhyloMap}
\title{Phylomap of Brassica nigra}
\format{
A tibble with 47,953 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from brassicadb}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Brassica_nigra.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from brassicadb (Brani_San_V1.1).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
