% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brassica_rapa.R
\docType{data}
\name{Brassica_rapa.PhyloMap}
\alias{Brassica_rapa.PhyloMap}
\title{Phylomap of Brassica rapa}
\format{
A tibble with 46,250 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from brassicadb}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Brassica_rapa.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from brassicadb (Brara_Chiifu_V3.0).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
