% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Caenorhabditis_elegans_WormBase_WS230.R
\docType{data}
\name{Caenorhabditis_elegans_WormBase_WS230.PhyloMap}
\alias{Caenorhabditis_elegans_WormBase_WS230.PhyloMap}
\title{Phylomap of Caenorhabditis elegans WormBase WS230}
\format{
A tibble with 25,634 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from WormBase}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Caenorhabditis_elegans_WormBase_WS230.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from WormBase (WS230 release).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
