% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coprinopsis_cinerea.R
\docType{data}
\name{Coprinopsis_cinerea.PhyloMap}
\alias{Coprinopsis_cinerea.PhyloMap}
\title{Phylomap of Coprinopsis cinerea}
\format{
A tibble with 13,393 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from BroadInstitutev3.0 assembly}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Coprinopsis_cinerea.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from BroadInstitutev3.0 assembly (https://www.broadinstitute.org/fungal-genome-initiative/coprinopsis-cinerea-genome-project).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
