% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coprinopsis_cinerea_JGI_Copci_AmutBmut1.R
\docType{data}
\name{Coprinopsis_cinerea_JGI_Copci_AmutBmut1.PhyloMap}
\alias{Coprinopsis_cinerea_JGI_Copci_AmutBmut1.PhyloMap}
\title{Phylomap of Coprinopsis cinerea JGI Copci AmutBmut1}
\format{
A tibble with 14,242 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from JGI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Coprinopsis_cinerea_JGI_Copci_AmutBmut1.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from JGI proteomes (Coprinopsis cinerea AmutBmut pab1-1 v1.0 (Copci_AmutBmut1) JGI Project Id: 1004746).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
