% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Danio_rerio.R
\docType{data}
\name{Danio_rerio.PhyloMap}
\alias{Danio_rerio.PhyloMap}
\title{Phylomap of Danio rerio}
\format{
A tibble with 43,153 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{ProteinID annotation from ENSEMBL}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Danio_rerio.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from ENSEMBL proteomes (Zv9).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
