% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drosophila_melanogaster_ENSEMBL_BDGP5.R
\docType{data}
\name{Drosophila_melanogaster_ENSEMBL_BDGP5.PhyloMap}
\alias{Drosophila_melanogaster_ENSEMBL_BDGP5.PhyloMap}
\title{Phylomap of Drosophila melanogaster ENSEMBL BDGP5}
\format{
A tibble with 26,950 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{ProteinID annotation from ENSEMBL}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Drosophila_melanogaster_ENSEMBL_BDGP5.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from ENSEMBL proteomes (BDGP5).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
