% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gallus_gallus.R
\docType{data}
\name{Gallus_gallus.PhyloMap}
\alias{Gallus_gallus.PhyloMap}
\title{Phylomap of Gallus gallus}
\format{
A tibble with 16,354 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{ProteinID annotation from ENSEMBL}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Gallus_gallus.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from ENSEMBL proteomes (Galgal4).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
