% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homo_sapiens.R
\docType{data}
\name{Homo_sapiens.PhyloMap}
\alias{Homo_sapiens.PhyloMap}
\title{Phylomap of Homo sapiens}
\format{
A tibble with 20,598 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{chr GeneID annotation from UniProt}
}
}
\source{
Lotharukpong (2023) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Homo_sapiens.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from Uniprot proteomes.
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
