% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homo_sapiens_ENSEMBL_GRCh37hg19.R
\docType{data}
\name{Homo_sapiens_ENSEMBL_GRCh37hg19.PhyloMap}
\alias{Homo_sapiens_ENSEMBL_GRCh37hg19.PhyloMap}
\title{Phylomap of Homo sapiens ENSEMBL GRCh37hg19}
\format{
A tibble with 104,761 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{ProteinID annotation from ENSEMBL}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Homo_sapiens_ENSEMBL_GRCh37hg19.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from ENSEMBL proteomes (GRCh37(hg19)).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
