% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypsibius_dujardini.R
\docType{data}
\name{Hypsibius_dujardini.PhyloMap}
\alias{Hypsibius_dujardini.PhyloMap}
\title{Phylomap of Hypsibius dujardini Hduj_transcriptome_v1}
\format{
A tibble with 16,747 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{transcriptID annotation from Hduj_transcriptome_v1 assembly + Transdecoder}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Hypsibius_dujardini.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference transcript sequences from Hduj_transcriptome_v1 assembly (https://www.ncbi.nlm.nih.gov/nuccore?linkname=bioproject_nuccore_transcript&from_uid=271450). \href{https://github.com/TransDecoder/TransDecoder}{Transdecoder} -m 50 was performed to obtain protein sequences.
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
