% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lentinus_tigrinus.R
\docType{data}
\name{Lentinus_tigrinus.PhyloMap}
\alias{Lentinus_tigrinus.PhyloMap}
\title{Phylomap of Lentinus tigrinus}
\format{
A tibble with 15,581 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from JGI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Lentinus_tigrinus.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from JGI proteomes (Lentinus tigrinus ALCF2SS1-6 v1.0 (Lenti6_1) JGI project ID: 1020066).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
