% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monodelphis_domestica.R
\docType{data}
\name{Monodelphis_domestica.PhyloMap}
\alias{Monodelphis_domestica.PhyloMap}
\title{Phylomap of Monodelphis domestica}
\format{
A tibble with 22,310 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{ProteinID annotation from ENSEMBL}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Monodelphis_domestica.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from ENSEMBL proteomes (BROADO5).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
