% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mortierella_elongata.R
\docType{data}
\name{Mortierella_elongata.PhyloMap}
\alias{Mortierella_elongata.PhyloMap}
\title{Phylomap of Mortierella elongata}
\format{
A tibble with 14,969 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{chr GeneID annotation from UniProt}
}
}
\source{
Manley et al. (2023) \emph{G3 Genes|Genomes|Genetics}
\url{https://doi.org/10.1093/g3journal/jkad077}
}
\usage{
Mortierella_elongata.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on refence protein sequences from Uniprot proteomes.
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{ultra-sensitive mode}.
}
\keyword{datasets}
