% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mus_musculus_ENSEMBL_GRC38mm10.R
\docType{data}
\name{Mus_musculus_ENSEMBL_GRC38mm10.PhyloMap}
\alias{Mus_musculus_ENSEMBL_GRC38mm10.PhyloMap}
\title{Phylomap of Mus musculus ENSEMBL GRC38mm10}
\format{
A tibble with 68,379 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{ProteinID annotation from ENSEMBL}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Mus_musculus_ENSEMBL_GRC38mm10.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from ENSEMBL proteomes (GRC38(mm10)).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
