% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nematostella_vectensis_NvT1.R
\docType{data}
\name{Nematostella_vectensis_NvT1.PhyloMap}
\alias{Nematostella_vectensis_NvT1.PhyloMap}
\title{Phylomap of Nematostella vectensis NvT1}
\format{
A tibble with 102,739 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{transcriptID annotation from NvT1 assembly + Transdecoder}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Nematostella_vectensis_NvT1.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference transcript sequences from NvT1 assembly (https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71086). \href{https://github.com/TransDecoder/TransDecoder}{Transdecoder} -m 50 was performed to obtain protein sequences.
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
