% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phanerodontia_chrysosporium.R
\docType{data}
\name{Phanerodontia_chrysosporium.PhyloMap}
\alias{Phanerodontia_chrysosporium.PhyloMap}
\title{Phylomap of Phanerodontia chrysosporium}
\format{
A tibble with 13,602 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from JGI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Phanerodontia_chrysosporium.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from JGI proteomes (Phanerochaete chrysosporium RP-78 v2.2).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
