% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phycomyces_blakesleeanus.R
\docType{data}
\name{Phycomyces_blakesleeanus.PhyloMap}
\alias{Phycomyces_blakesleeanus.PhyloMap}
\title{Phylomap of Phycomyces blakesleeanus}
\format{
A tibble with 16,528 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{chr GeneID annotation from UniProt}
}
}
\source{
Manley et al. (2023) \emph{G3 Genes|Genomes|Genetics}
\url{https://doi.org/10.1093/g3journal/jkad077}
}
\usage{
Phycomyces_blakesleeanus.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on refence protein sequences from Uniprot proteomes.
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{ultra-sensitive mode}.
}
\keyword{datasets}
