% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomap.R
\name{phylomap}
\alias{phylomap}
\title{Turns a phylogeny into a graph projected onto geographical space}
\usage{
phylomap(tree, data, crs = 4326)
}
\arguments{
\item{tree}{Phylogenetic tree of class `phylo`.}

\item{data}{An object of class `data.frame` or `matrix` with coordinates of the tips in the tree in columns 'lon' and 'lat' and rownames = tip labels}

\item{crs}{The coordinate reference system of the lon and lat data. Default is 4326, the EPSG code for World Geodetic System 1984. See https://epsg.io/4326.}
}
\value{
An `sfnetwork` object.
}
\description{
This function takes a phylogeny object (class phylo) and the geographical
coordinates of the leaves, infers the internal node locations, and then
converts the phylogeny into an sfnetworks spatial graph.
}
\examples{
library(phylomapR)
library(ape)
library(phytools)
library(ggraph)
library(ggplot2)
library(magrittr)
library(dplyr)
library(sfnetworks)
library(tidygraph)
random_tree_rooted<-rtree(26,rooted = T,tip.label = letters)
random_xy<-data.frame(
  lon=fastBM(random_tree_rooted,bounds=c(15,25),sig2=2),
  lat=fastBM(random_tree_rooted,bounds=c(0,10),sig2=2)
)
# Note that fastBM produces named vectors so we don't have to add rownames ourselves
test_rooted<-phylomap(random_tree_rooted,random_xy)
### basic plot ----
test_rooted \%>\%
  ggraph('sf')+
  geom_node_sf()+
  geom_edge_sf()+
  theme_bw()

### add graphy info to nodes and edges and plot ----
test_rooted \%>\%
  activate(nodes) \%>\%
  mutate(centrality=centrality_betweenness(normalized = F)) \%>\%
  activate(edges) \%>\%
  mutate(betweenness=centrality_edge_betweenness()) \%>\%
  ggraph('sf')+
  geom_edge_sf(aes(alpha=betweenness),linewidth=2)+
  geom_node_sf(aes(colour=centrality,size=centrality))+
  theme_bw()+
  scale_colour_viridis_c()

}
