% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomapr-package.R
\docType{package}
\name{phylomapr}
\alias{phylomapr}
\title{Quick retrieval of phylomaps}
\source{
Lotharukpong (2023) \emph{Unpublished}
\url{https://lotharukpongjs.github.io/phylomapr/}
}
\description{
Gene founder events facilitate evolutionary innovations. phylomapr enables quick retrieval of precomputed gene age maps (phylomaps) in R. Gene age maps loaded from phylomapr integrate seamlessly with myTAI.
}
\details{
phylomapr
}
