% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Platynereis_dumerilii.R
\docType{data}
\name{Platynereis_dumerilii.PhyloMap}
\alias{Platynereis_dumerilii.PhyloMap}
\title{Phylomap of Platynereis dumerilii Pdum_transcriptome_v1}
\format{
A tibble with 12,438 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{transcriptID annotation from Pdum_transcriptome_v1 assembly + Transdecoder}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Platynereis_dumerilii.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference transcript sequences from Pdum_transcriptome_v1 assembly (https://www.ncbi.nlm.nih.gov/nuccore?linkname=bioproject_nuccore_transcript&from_uid=271451). \href{https://github.com/TransDecoder/TransDecoder}{Transdecoder} -m 50 was performed to obtain protein sequences.
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
