% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pleurotus_ostreatus.R
\docType{data}
\name{Pleurotus_ostreatus.PhyloMap}
\alias{Pleurotus_ostreatus.PhyloMap}
\title{Phylomap of Pleurotus ostreatus}
\format{
A tibble with 12,330 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from JGI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Pleurotus_ostreatus.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from JGI proteomes (Pleurotus ostreatus PC15 v2.0 (PleosPC15_2)).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
