% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pterula_gracilis.R
\docType{data}
\name{Pterula_gracilis.PhyloMap}
\alias{Pterula_gracilis.PhyloMap}
\title{Phylomap of Pterula gracilis}
\format{
A tibble with 12,873 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from JGI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Pterula_gracilis.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from JGI proteomes (Pterula gracilis CBS309.79 v1.0 (Ptegra1)).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
