% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomyces_spectabilis.R
\docType{data}
\name{Radiomyces_spectabilis.PhyloMap}
\alias{Radiomyces_spectabilis.PhyloMap}
\title{Phylomap of Radiomyces spectabilis}
\format{
A tibble with 10,905 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{chr GeneID annotation from UniProt}
}
}
\source{
Manley et al. (2023) \emph{G3 Genes|Genomes|Genetics}
\url{https://doi.org/10.1093/g3journal/jkad077}
}
\usage{
Radiomyces_spectabilis.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from Uniprot proteomes.
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{ultra-sensitive mode}.
}
\keyword{datasets}
