% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rattus_norvegicus.R
\docType{data}
\name{Rattus_norvegicus.PhyloMap}
\alias{Rattus_norvegicus.PhyloMap}
\title{Phylomap of Rattus norvegicus}
\format{
A tibble with 26,153 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{ProteinID annotation from ENSEMBL}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Rattus_norvegicus.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from ENSEMBL proteomes (Rnor_5.0).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
