% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhizophagus_irregularis.R
\docType{data}
\name{Rhizophagus_irregularis.PhyloMap}
\alias{Rhizophagus_irregularis.PhyloMap}
\title{Phylomap of Rhizophagus irregularis}
\format{
A tibble with 31,217 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{chr GeneID annotation from UniProt}
}
}
\source{
Manley et al. (2023) \emph{G3 Genes|Genomes|Genetics}
\url{https://doi.org/10.1093/g3journal/jkad077}
}
\usage{
Rhizophagus_irregularis.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from Uniprot proteomes.
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{ultra-sensitive mode}.
}
\keyword{datasets}
