% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rickenella_mellea.R
\docType{data}
\name{Rickenella_mellea.PhyloMap}
\alias{Rickenella_mellea.PhyloMap}
\title{Phylomap of Rickenella mellea}
\format{
A tibble with 17,139 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from JGI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Rickenella_mellea.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from JGI proteomes (Rickenella mellea v1.0 SZMC22713 (Ricmel1), JGI project ID:1075324).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
