% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schizophyllum_commune.R
\docType{data}
\name{Schizophyllum_commune.PhyloMap}
\alias{Schizophyllum_commune.PhyloMap}
\title{Phylomap of Schizophyllum commune}
\format{
A tibble with 16,319 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{GeneID annotation from JGI}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Schizophyllum_commune.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from JGI proteomes (Schizophyllum commune H4-8 v3.0 (Schco3), JGI project ID: 16355).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
