% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strongylocentrotus_purpuratus.R
\docType{data}
\name{Strongylocentrotus_purpuratus.PhyloMap}
\alias{Strongylocentrotus_purpuratus.PhyloMap}
\title{Phylomap of Strongylocentrotus purpuratus}
\format{
A tibble with 38,475 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{proteinID annotation from NCBI}
}
}
\source{
Bermejo (2023) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Strongylocentrotus_purpuratus.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference protein sequences from NCBI protein coding genes (https://www.ncbi.nlm.nih.gov/datasets/gene/taxon/7668/?gene-type=Protein-coding).
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{ultra-sensitive mode}.
}
\keyword{datasets}
