% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strongylocentrotus_purpuratus_WHL22.R
\docType{data}
\name{Strongylocentrotus_purpuratus_WHL22.PhyloMap}
\alias{Strongylocentrotus_purpuratus_WHL22.PhyloMap}
\title{Phylomap of Strongylocentrotus purpuratus WHL22}
\format{
A tibble with 35,907 rows and 2 variables:
\describe{
\item{Phylostratum}{dbl Phylostratum (or gene age) assignment}
\item{GeneID}{transcriptID annotation from WHL22 assembly + Transdecoder}
}
}
\source{
Kalábová (2024) \emph{Unpublished}
\url{https://github.com/LotharukpongJS/phylomapr}
}
\usage{
Strongylocentrotus_purpuratus_WHL22.PhyloMap
}
\description{
Gene ages inferred using \href{https://github.com/josuebarrera/GenEra}{GenEra} on reference transcript sequences from WHL22 assembly (https://www.ncbi.nlm.nih.gov/nuccore/1583097974). \href{https://github.com/TransDecoder/TransDecoder}{Transdecoder} -m 50 was performed to obtain protein sequences.
Note: \href{https://github.com/bbuchfink/diamond}{DIAMOND} was run using the \verb{sensitive mode}.
}
\keyword{datasets}
